+++ apps/minecraft/next.config.js
@@ -1,0 +1,8 @@
+const withTM = require("next-transpile-modules")(["three"]);
+
+module.exports = withTM({
+  reactStrictMode: true,
+  future: {
+    webpack5: true,
+  },
+});