+++ apps/minecraft/src/pages/_app.tsx
@@ -1,0 +1,16 @@
+import React, { FC } from "react";
+import Head from "next/head";
+import type { AppProps } from "next/app";
+
+const App: FC<AppProps> = ({ Component, pageProps }) => (
+  <>
+    <Head>
+      <title>React Minecraft</title>
+      <meta charSet="utf-8" />
+      <meta name="viewport" content="initial-scale=1.0, width=device-width" />
+    </Head>
+    <Component {...pageProps} />
+  </>
+);
+
+export default App;